requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.11
Flask==3.1.0
flask-cors==5.0.1
flatdict==4.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
loguru==0.7.3
MarkupSafe==3.0.2
motor==3.7.0
numpy>=1.21.0,<2.0.0
ollama==0.4.7
openai==1.65.3
pandas==2.2.3
psutil==7.0.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
torch>=2.0.0,<2.3.0
transformers>=4.30.0,<5.0.0
oss2>=2.15.0
Pillow>=8.0.0
